projects/wml-table/src/lib/wml-table-zero/wml-table-zero-sample-row/wml-table-zero-sample-row.component.spec.ts
// testing
import { ComponentFixture } from '@angular/core/testing';
import { wmlTestUtils } from '../test-utils/test-utils';
// rxjs
import { Subject } from 'rxjs';

import { WMLTableZeroSampleRowComponent } from './wml-table-zero-sample-row.component';


describe('WMLTableZeroSampleRowComponent', () => {
  let cpnt: WMLTableZeroSampleRowComponent;
  let fixture: ComponentFixture<WMLTableZeroSampleRowComponent>;

  beforeEach(async () => {



    await wmlTestUtils.configureTestingModuleForComponents(WMLTableZeroSampleRowComponent);


    ({fixture, cpnt} = wmlTestUtils.grabComponentInstance(WMLTableZeroSampleRowComponent));
    fixture.detectChanges()
  })

  describe("init", () => {

    it("should create", () => {
      expect(cpnt).toBeTruthy()
    })

    it("should have all values initalize properly", () => {
      expect(cpnt.myClass).toEqual('WMLTableZeroSampleRowView')
    })

    it("should have all properties be the correct class instance", () => {
      expect(cpnt.ngUnsub).toBeInstanceOf(Subject<void>)
    })
  })

  describe("ngOnDestroy",()=>{

    beforeEach(()=>{
      spyOn(cpnt.ngUnsub,'next')
      spyOn(cpnt.ngUnsub,'complete')
    })

    it(` when called |
     as appropriate |
     does the required action `,()=>{
        // act
        cpnt.ngOnDestroy();

        // assert
        expect(cpnt.ngUnsub.next).toHaveBeenCalled();
        expect(cpnt.ngUnsub.complete).toHaveBeenCalled();
    })
  })
});
